GAL_005 changed MainActivityEspressoTest [FM,GD]
